requirements.txt
Flask~=2.2.2
Boto3==1.26.29
python-dotenv~=0.21.0
Werkzeug~=2.2.2
Pillow~=9.3.0